requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1